Modify for usage cluster pruning

+++ InformationRetrieval/Index/Document.cs
@@ -7,7 +7,7 @@ namespace InformationRetrieval.Index
     public class Document
     {
         public string Filename { get; set; } = "";
-        public int Length { get; set; }
+        public double Length { get; set; }
         public List<Term> Terms { get; set; } = new List<Term>();
         public Dictionary<int, double> Correlation { get; set; } = new Dictionary<int, double>();
         public int[] DocVec;
